vim-patch:9.1.0649: Wrong comment for "len" argument of call_simple_func()
Problem:  Wrong comment for "len" argument of call_simple_func().
Solution: Remove the "or -1 to use strlen()".  Also change its type to
          size_t to remove one cast. (zeertzjq)

closes: vim/vim#15410

https://github.com/vim/vim/commit/c1ed788c1b41db9b5f1ef548dc877f771f535bbe

+++ src/nvim/eval.c
@@ -2630,7 +2630,7 @@ static int may_call_simple_func(char *arg, typval_T *rettv)
   if (parens != NULL && *skipwhite(parens + 2) == NUL) {
     char *p = strncmp(arg, "<SNR>", 5) == 0 ? skipdigits(arg + 5) : arg;
     if (to_name_end(p, true) == parens) {
-      r = call_simple_func(arg, (int)(parens - arg), rettv);
+      r = call_simple_func(arg, (size_t)(parens - arg), rettv);
     }
   }
   return r;
